Changes performed: * Fix for resolving IUserValidator interface inside bootstrapper * Remove WriteConfig method from AuthenticatorPlugin.cs and move it to the specific plugin(WSP) * Create authenticator properties specific for WSP plugin
